countdown

+++ app/src/main/java/com/example/wetterbericht/data/local/entity/habits/HabitsLocal.kt
@@ -15,7 +15,7 @@ data class HabitsLocal(
     @NonNull
     val title : String,
     @NonNull
-    val minuteFocus : Int,
+    val minuteFocus : Long,
     @NonNull
     val startTime : String,
     @NonNull
