+++ app/src/main/java/com/example/wetterbericht/presentation/fragment/habits/insert/InsertDetailHabitsFragment.kt
@@ -50,7 +50,7 @@ class InsertDetailHabitsFragment : Fragment() {
         val insertData = HabitsLocal(
             0,
             name,
-            duration,
+            duration.toLong(),
             "0",
             ""
         )
